--- src/code42cli/cmds/enums.py
@@ -1,8 +1,0 @@
-class OutputFormat:
-    TABLE = "TABLE"
-    CSV = "CSV"
-    JSON = "JSON"
-    RAW = "RAW-JSON"
-
-    def __iter__(self):
-        return iter([self.TABLE, self.CSV, self.JSON, self.RAW])
